+++ prio_utils/circuit.go
@@ -1,4 +1,4 @@
-package utils
+package prio_utils
 
 import (
 	"github.com/henrycg/prio/circuit"
