+++ prio_utils/int.go
@@ -1,4 +1,4 @@
-package utils
+package prio_utils
 
 import (
 "log"
