+++ prio_utils/rand.go
@@ -1,4 +1,4 @@
-package utils
+package prio_utils
 
 import (
 	"bufio"
