+++ prio_utils/utils_func.go
@@ -1,4 +1,4 @@
-package utils
+package prio_utils
 
 import (
 	"math/big"
