--- utils/mpc.go
@@ -1,2 +1,0 @@
-package utils
-
